--- converage
@@ -1,8 +1,0 @@
-mode: set
-/home/nicte/Escritorio/Sa/practica-sa-201404104/main.go:12.14,17.12 2 0
-/home/nicte/Escritorio/Sa/practica-sa-201404104/main.go:22.2,22.12 1 0
-/home/nicte/Escritorio/Sa/practica-sa-201404104/main.go:28.2,28.12 1 0
-/home/nicte/Escritorio/Sa/practica-sa-201404104/main.go:35.2,36.31 2 0
-/home/nicte/Escritorio/Sa/practica-sa-201404104/main.go:17.12,20.3 2 0
-/home/nicte/Escritorio/Sa/practica-sa-201404104/main.go:22.12,25.3 2 0
-/home/nicte/Escritorio/Sa/practica-sa-201404104/main.go:28.12,32.3 2 0
