--- run.sh
@@ -1,3 +1,0 @@
-go test -coverprofile=converage ; 
-sonar-scanner 
-
