+++ serivicio-cliente/Cliente.go
@@ -1,14 +1,13 @@
-package servicio_cliente
+package main
 
 import (
-	"../utils"
+	"./utils"
 	"./peticiones-cliente"
 	"encoding/json"
 	"fmt"
 	"github.com/gorilla/mux"
 	"math/rand"
 	"net/http"
-
 )
 
 
@@ -91,16 +90,7 @@ func getEstadoRepartidor(w http.ResponseWriter, r *http.Request)  {
 }
 
 
-func Handle() {
-
-	utils.RegistrarServicio(&utils.ServicioData{"8080", "solicitar_pedido","/solicitar_pedido","cliente","POST"}, "POST","8085","/registrar_microservicio")
-	utils.RegistrarServicio(&utils.ServicioData{"8080", "get_estado_restaurante","/get_estado_restaurante","cliente","GET"}, "POST","8085","/registrar_microservicio")
-	utils.RegistrarServicio(&utils.ServicioData{"8080", "get_estado_repartidor","/get_estado_repartidor","cliente","GET"}, "POST","8085","/registrar_microservicio")
-
-	fmt.Println("Escuhando puerto 8080")
-	peticiones_cliente.HashPedido = make(map[int]int)
-	peticiones_cliente.Codigo  = -1
-	peticiones_cliente.IdPedido = -1
+func handle() {
 
 	router := mux.NewRouter()
 	router.HandleFunc("/solicitar_pedido",solicitarPedido).Methods("POST")
@@ -111,3 +101,19 @@ func Handle() {
 
 
 
+
+
+func main()  {
+
+	utils.RegistrarServicio(&utils.ServicioData{"8080", "solicitar_pedido","/solicitar_pedido","cliente","POST"}, "POST","8085","/registrar_microservicio")
+	utils.RegistrarServicio(&utils.ServicioData{"8080", "get_estado_restaurante","/get_estado_restaurante","cliente","GET"}, "POST","8085","/registrar_microservicio")
+	utils.RegistrarServicio(&utils.ServicioData{"8080", "get_estado_repartidor","/get_estado_repartidor","cliente","GET"}, "POST","8085","/registrar_microservicio")
+
+	fmt.Println("Escuhando puerto 8080")
+	peticiones_cliente.HashPedido = make(map[int]int)
+	peticiones_cliente.Codigo  = -1
+	peticiones_cliente.IdPedido = -1
+	handle()
+
+
+}
