+++ servicio-repartidor/Repartidor.go
@@ -1,11 +1,11 @@
-package servicio_repartidor
+package main
 
 import (
 	"encoding/json"
 	"fmt"
 	"github.com/gorilla/mux"
 	"net/http"
-	"../utils"
+	"./utils"
 )
 
 
@@ -139,12 +139,7 @@ func marcarPedido(w http.ResponseWriter, r *http.Request)  {
 }
 
 //funcion que expone los servicios del repartidor
-func Handle()  {
-
-	utils.RegistrarServicio(&utils.ServicioData{"8082", "recibir_pedidio","/recibir_pedidio","repartidor","POST"}, "POST","8085","/registrar_microservicio")
-	utils.RegistrarServicio(&utils.ServicioData{"8082", "informar_estado_cliente","/informar_estado_cliente","repartidor","GET"}, "POST","8085","/registrar_microservicio")
-	utils.RegistrarServicio(&utils.ServicioData{"8082", "marcar_pedido","/marcar_pedido","repartidor","POST"}, "POST","8085","/registrar_microservicio")
-
+func handle()  {
 
 	router := mux.NewRouter()
 	router.HandleFunc("/recibir_pedidio",recibirPedido).Methods("POST")
@@ -154,4 +149,13 @@ func Handle()  {
 
 }
 
+func main (){
 
+	utils.RegistrarServicio(&utils.ServicioData{"8082", "recibir_pedidio","/recibir_pedidio","repartidor","POST"}, "POST","8085","/registrar_microservicio")
+	utils.RegistrarServicio(&utils.ServicioData{"8082", "informar_estado_cliente","/informar_estado_cliente","repartidor","GET"}, "POST","8085","/registrar_microservicio")
+	utils.RegistrarServicio(&utils.ServicioData{"8082", "marcar_pedido","/marcar_pedido","repartidor","POST"}, "POST","8085","/registrar_microservicio")
+
+	println("escuchando el puerto 8082")
+	handle()
+
+}
