--- sonar-project.properties
@@ -1,15 +1,0 @@
-sonar.projectKey=sa-practica
-sonar.projectName=sa-practica
-sonar.projectVersion=1.0
-# GoLint report path, default value is report.xml 
-sonar.golint.reportPath=report.xml 
-# Cobertura like coverage report path, default value is coverage.xml 
-sonar.coverage.reportPath=coverage.xml 
-# if you want disabled the DTD verification for a proxy problem for example, true by default 
-sonar.coverage.dtdVerification=true
-# JUnit like test report, default value is test.xml
-sonar.test.reportPath=test.xml 
-sonar.sources=./
-sonar.tests=./
-sonar.test.inclusions=**/**_test.go
-sonar.sources.inclusions=**/**.go
